requirements.txt
alabaster==0.7.12
asgiref==3.5.2
Babel==2.10.3
certifi==2022.6.15
charset-normalizer==2.1.1
colorama==0.4.5
defusedxml==0.7.1
Django==4.1
django-environ==0.9.0
docutils==0.19
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
imagesize==1.4.1
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pguard==0.4.0
Pillow==9.2.0
Pygments==2.13.0
pyparsing==3.0.9
python-http-client==3.3.7
pytz==2022.2.1
requests==2.28.1
sendgrid==6.9.7
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==5.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
starkbank-ecdsa==2.1.0
tzdata==2022.2
urllib3==1.26.12
